feat: move to shared component
