+++ frontend/src/scenes/onboarding/error-tracking/source-maps/automated-technologies/NextJSSourceMapsInstructions.tsx
@@ -1,42 +1,19 @@
-import { useActions, useValues } from 'kea'
-
-import { LemonBanner, LemonButton } from '@posthog/lemon-ui'
+import { useValues } from 'kea'
 
 import { CodeSnippet, Language } from 'lib/components/CodeSnippet'
-import { API_KEY_SCOPE_PRESETS } from 'lib/scopes'
 import { apiHostOrigin } from 'lib/utils/apiHost'
-import { personalAPIKeysLogic } from 'scenes/settings/user/personalAPIKeysLogic'
 import { teamLogic } from 'scenes/teamLogic'
 
+import { SourceMapsAPIKeyBanner } from '../SourceMapsAPIKeyBanner'
 import { VerifySourceMaps } from '../VerifySourceMaps'
 
 export function NextJSSourceMapsInstructions(): JSX.Element {
     const { currentTeam } = useValues(teamLogic)
-    const { setEditingKeyId, setEditingKeyValues } = useActions(personalAPIKeysLogic)
     const host = apiHostOrigin()
-
-    const openAPIKeyModal = (): void => {
-        const preset = API_KEY_SCOPE_PRESETS.find((p) => p.value === 'source_map_upload')
-        if (preset) {
-            setEditingKeyId('new')
-            setEditingKeyValues({
-                preset: preset.value,
-                label: preset.label,
-                scopes: preset.scopes,
-                access_type: preset.access_type || 'all',
-            })
-        }
-    }
 
     return (
         <>
-            <LemonBanner type="info" className="mb-4">
-                <strong>Note:</strong> The project API key is not the same as the personal API token required to upload
-                source maps.{' '}
-                <LemonButton type="secondary" size="xsmall" onClick={openAPIKeyModal}>
-                    Generate personal API key
-                </LemonButton>
-            </LemonBanner>
+            <SourceMapsAPIKeyBanner />
 
             <h3>Install the PostHog Next.js config package</h3>
             <p>
